primera etapa de manejo de planes

+++ CommerceCore.ML/Security/SecurityOption.cs
@@ -8,14 +8,14 @@ namespace CommerceCore.ML.Security
 {
     public class SecurityOption
     {
-        public long id { get; set; }
-        public byte menu { get; set; }
+        public int id { get; set; }
+        public int menu { get; set; }
         public int? grouper { get; set; }
         public string name { get; set; }
         public string text { get; set; }
         public string pathIcon { get; set; }
         public string url { get; set; }
-        public byte orderShow { get; set; }
+        public int orderShow { get; set; }
         public bool active { get; set; }
     }
 }
